#48 Exchanged spacetime with a custom version
This change only is temporary because of missinng type definitions for spacetime. For more details please see https://github.com/spencermountain/spacetime/issues/255. A pull request to spacetime has been opened. Once this is merged and the fix is published we can switch back. I´ll open an issue for this later.